fix MAX_PATH_LEN
The `MAX_PATH_LEN` **MUST** be reduced by 39 from 99 to 60. Why? Because
(and forgot this) the V7 format, which is the default for many systems,
only supports paths up to 99 characters. If a file name exceeds 99,
then tar will not include it an exit non-zero.

The 39 is because a leading "UUID-slot_num/" such as
`12345678-1234-4321-abcd-1234567890ab-5/" is 39 characters.

While we worked on the tar standard, we **FORGOT** (sorry tm Canada 🇨🇦)
that the v7 tar had this limitation. This bug was there for years making
the effective maximum path to be 60 characters. Nobody complained about
not being able to have a path between 61 and 99 characters before,
so reducing MAX_PATH_LEN to 60 shouldn't impact very many people.

Temporarily removed 3 bad answer files so that `make release`,
the GitHub actions, and ``test_ioccc/mkiocccentry_slots.sh` (via `make test`)
will pass while we test the above, **AND** while we work on the
fix for issue #1359.

Changed `MKIOCCCENTRY_REPO_VERSION` to "2.9.5 2025-11-14".

NOTE: Changed the following due to `MAX_PATH_LEN` change as noted above.
Changed `MKIOCCCENTRY_VERSION` to "2.1.10 2025-11-14".
Changed `FNAMCHK_VERSION` to "2.1.6 2025-11-14".
Changed `TXZCHK_VERSION` to "2.0.13 2025-11-14".
Changed `CHKENTRY_VERSION` to "2.1.11 2025-11-14".
Changed `CHKSUBMIT_VERSION` to "1.0.2 2025-11-14".

Ran `make release` to test the above.

TODO: Fix issue #1359: working on a fix now.