Fix bug preventing .. in workdir
Fix error preventing .. in workdir - issue #1367. This required
changing a boolean in the call of canon_path() in both txzchk and
fnamchk.

NOTE: this will NOT create a tarball with .. in the path. It just
allows one to have their workdir to have in its path ...

Updated TXZCHK_VERSION to "2.1.1 2025-11-21".
Updated FNAMCHK_VERSION to "2.2.1 2025-11-21".