Typo fix in txzchk
Changed MIN_TXZCHK_VERSION to "2.1.0 2025-11-18".
Updated TXZCHK_VERSION to "2.1.1 2025-12-20"`.

IMPORTANT NOTES: this is NOT a functional change! It is NOT necessary to
have this before IOCCC30 and this is one of the reasons I changed the
minimum version to what is now the previous version of txzchk. This is
only done because it annoys me - NOT because it's important to anyone
but me (and it's certainly unimportant for the contest). In other words
even if this is not merged (or if it is not installed on the submit
server) until after IOCCC29 it does not matter in the slightest. There
is no real need other than to make me happier about the typo being
killed.

... in fact although it has no functional change and although the min
version should be fine I would personally not install this on the server
until the next time - but that's me. I leave that to Landon nonetheless.